+++ skoolkit/bin2sna.py
@@ -1,4 +1,4 @@
-# Copyright 2016-2018, 2020 Richard Dymond (rjdymond@gmail.com)
+# Copyright 2016-2018, 2020, 2023 Richard Dymond (rjdymond@gmail.com)
 #
 # This file is part of SkoolKit.
 #
@@ -18,7 +18,7 @@ import os
 import argparse
 
 from skoolkit import integer, read_bin_file, VERSION
-from skoolkit.snapshot import poke, print_reg_help, print_state_help, write_z80v3
+from skoolkit.snapshot import poke, print_reg_help, print_state_help, write_snapshot
 
 def run(infile, outfile, options):
     ram = list(read_bin_file(infile, 49152))
@@ -39,14 +39,14 @@ def run(infile, outfile, options):
         os.makedirs(parent_dir)
     registers = ['sp={}'.format(stack), 'pc={}'.format(start)] + options.reg
     state = ['border={}'.format(options.border)] + options.state
-    write_z80v3(outfile, snapshot[16384:], registers, state)
+    write_snapshot(outfile, snapshot[16384:], registers, state)
 
 def main(args):
     parser = argparse.ArgumentParser(
-        usage='bin2sna.py [options] file.bin [file.z80]',
-        description="Convert a binary (raw memory) file into a Z80 snapshot. "
+        usage='bin2sna.py [options] file.bin [OUTFILE]',
+        description="Convert a binary (raw memory) file into an SZX or Z80 snapshot. "
                     "'file.bin' may be a regular file, or '-' for standard input. "
-                    "If 'file.z80' is not given, it defaults to the name of the input file with '.bin' replaced by '.z80', "
+                    "If 'OUTFILE' is not given, it defaults to the name of the input file with '.bin' replaced by '.z80', "
                     "or 'program.z80' if reading from standard input.",
         add_help=False
     )