+++ skoolkit/tap2sna.py
@@ -33,7 +33,7 @@ from skoolkit.loadtracer import LoadTracer, get_edges
 from skoolkit.pagingtracer import Memory
 from skoolkit.simulator import (Simulator, A, F, B, C, D, E, H, L, IXh, IXl, IYh, IYl,
                                 SP, I, R, xA, xF, xB, xC, xD, xE, xH, xL, PC)
-from skoolkit.snapshot import FRAME_DURATIONS, move, poke, print_reg_help, print_state_help, write_z80v3
+from skoolkit.snapshot import FRAME_DURATIONS, move, poke, print_reg_help, print_state_help, write_snapshot
 
 SYSVARS = (
     255, 0, 0, 0,         # 23552 - KSTATE0
@@ -254,12 +254,12 @@ def get_tape_block_timings(first_byte, pause=3500000):
     # Data block
     return TapeBlockTimings(3223, 2168, (667, 735), 855, 1710, pause)
 
-def _write_z80(ram, options, fname):
+def _write_snapshot(ram, options, fname):
     parent_dir = os.path.dirname(fname)
     if parent_dir and not os.path.isdir(parent_dir):
         os.makedirs(parent_dir)
     write_line('Writing {0}'.format(fname))
-    write_z80v3(fname, ram, options.reg, options.state)
+    write_snapshot(fname, ram, options.reg, options.state)
 
 def _ram_operations(snapshot, ram_ops, blocks=None):
     counters = {}
@@ -962,7 +962,7 @@ Configure various properties of a simulated LOAD.
   Log to FILE all instructions executed during the simulated LOAD.
 """.strip())
 
-def make_z80(url, options, z80, config):
+def make_snapshot(url, options, outfile, config):
     tape_name, tape_type, tape = _get_tape(url, options.user_agent, options.tape_name)
     if options.tape_sum:
         md5sum = hashlib.md5(tape).hexdigest()
@@ -980,26 +980,26 @@ def make_z80(url, options, z80, config):
     else:
         blocks = [b[1] for b in tape_blocks]
         ram = _get_ram(blocks, options)
-    if z80 is None:
+    if outfile is None:
         if tape_name.lower().endswith(('.tap', '.tzx')):
             tape_name = tape_name[:-4]
-        z80 = tape_name + '.z80'
+        outfile = tape_name + '.z80'
     if options.output_dir:
-        z80 = os.path.join(options.output_dir, z80)
-    _write_z80(ram, options, z80)
+        outfile = os.path.join(options.output_dir, outfile)
+    _write_snapshot(ram, options, outfile)
 
 def main(args):
     config = get_config('tap2sna')
     parser = SkoolKitArgumentParser(
-        usage='\n  tap2sna.py [options] INPUT [snapshot.z80]\n  tap2sna.py @FILE [args]',
-        description="Convert a TAP or TZX file (which may be inside a zip archive) into a Z80 snapshot. "
+        usage='\n  tap2sna.py [options] INPUT [OUTFILE]\n  tap2sna.py @FILE [args]',
+        description="Convert a TAP or TZX file (which may be inside a zip archive) into an SZX or Z80 snapshot. "
                     "INPUT may be the full URL to a remote zip archive or TAP/TZX file, or the path to a local file. "
                     "Arguments may be read from FILE instead of (or as well as) being given on the command line.",
         fromfile_prefix_chars='@',
         add_help=False
     )
     parser.add_argument('url', help=argparse.SUPPRESS, nargs='?')
-    parser.add_argument('z80', help=argparse.SUPPRESS, nargs='?')
+    parser.add_argument('outfile', help=argparse.SUPPRESS, nargs='?')
     group = parser.add_argument_group('Options')
     group.add_argument('-c', '--sim-load-config', metavar='name=value', action='append', default=[],
                        help="Set the value of a simulated LOAD configuration parameter. "
@@ -1062,6 +1062,6 @@ def main(args):
         namespace.reg.append('pc={}'.format(namespace.start))
     update_options('tap2sna', namespace, namespace.params, config)
     try:
-        make_z80(namespace.url, namespace, namespace.z80, config)
+        make_snapshot(namespace.url, namespace, namespace.outfile, config)
     except Exception as e:
-        raise SkoolKitError("Error while getting snapshot {}: {}".format(os.path.basename(namespace.z80), e.args[0] if e.args else e))
+        raise SkoolKitError("Error while getting snapshot {}: {}".format(os.path.basename(namespace.outfile), e.args[0] if e.args else e))