+++ sphinx/source/changelog.rst
@@ -19,6 +19,8 @@ Changelog
 * Added support to :ref:`trace.py` for executing machine code in 128K snapshots
 * Added the ``--no-interrupts`` option to :ref:`trace.py` (to prevent the
   execution of interrupt routines, which are now executed by default)
+* Added support to :ref:`bin2sna.py`, :ref:`tap2sna.py` and :ref:`trace.py` for
+  writing SZX snapshots
 * Added support to :ref:`control directive loops <ctlLoops>` for avoiding
   repetition of an ``N`` directive at the start of a loop
 * :ref:`tapinfo.py` now shows the LINE number (if present) for 'Program:'